+++ jax/_src/numpy/util.py
@@ -399,3 +399,34 @@ def _where(condition, x=None, y=None):
   try: is_always_empty = core.is_empty_shape(np.shape(x))
   except: is_always_empty = False  # can fail with dynamic shapes
   return lax.select(condition, x, y) if not is_always_empty else x
+
+
+def _should_unpack_list_index(x):
+  """Helper for _eliminate_deprecated_list_indexing."""
+  return (isinstance(x, (np.ndarray, ndarray)) and np.ndim(x) != 0
+          or isinstance(x, (Sequence, slice))
+          or x is Ellipsis or x is None)
+
+
+def _eliminate_deprecated_list_indexing(idx):
+  # "Basic slicing is initiated if the selection object is a non-array,
+  # non-tuple sequence containing slice objects, [Ellipses, or newaxis
+  # objects]". Detects this and raises a TypeError.
+  if not isinstance(idx, tuple):
+    if isinstance(idx, Sequence) and not isinstance(idx, (ndarray, np.ndarray)):
+      # As of numpy 1.16, some non-tuple sequences of indices result in a warning, while
+      # others are converted to arrays, based on a set of somewhat convoluted heuristics
+      # (See https://github.com/numpy/numpy/blob/v1.19.2/numpy/core/src/multiarray/mapping.c#L179-L343)
+      # In JAX, we raise an informative TypeError for *all* non-tuple sequences.
+      if any(_should_unpack_list_index(i) for i in idx):
+        msg = ("Using a non-tuple sequence for multidimensional indexing is not allowed; "
+               "use `arr[tuple(seq)]` instead of `arr[seq]`. "
+               "See https://github.com/google/jax/issues/4564 for more information.")
+      else:
+        msg = ("Using a non-tuple sequence for multidimensional indexing is not allowed; "
+               "use `arr[array(seq)]` instead of `arr[seq]`. "
+               "See https://github.com/google/jax/issues/4564 for more information.")
+      raise TypeError(msg)
+    else:
+      idx = (idx,)
+  return idx